* Beginning to attempt to bring GAMS back into Windows compilability * Added capnp feature, similar to the one added to MADARA * Fixed warnings about conversions in several places * Updated base_build.bat script to be slightly better for windows building but still requires forcing installation of capnp or other tools. I may have to give up on .bat scripting and just use perl or something else here to properly set variables and dictate installation flow.

+++ src/gams/algorithms/FormationSync.cpp
@@ -442,8 +442,8 @@ gams::algorithms::FormationSync::generate_plan(int formation)
        * the square
        **/
 
-      int square_size =(int)2 * group_members_.size();
-      int num_rows =(size_t)std::sqrt((double)square_size);
+      int square_size = 2 * (int)group_members_.size();
+      int num_rows = (int)std::sqrt((double)square_size);
       int num_cols = num_rows;
       int base = 0;
       int row = 0;
